metoolok/requirements.txt
streamlit
aiohttp
requests
PyPDF2
python-dotenv